NCL-748: Addresses most of the warnings that come from compiling with -Wall
There are still a few more difficult cases to resolve. Also I am leaving all
files that are used to implement HDF5 support for Wei to handle as he is 
more familiar with that code.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11763 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/ncl/FileSupport.c
@@ -1490,11 +1490,11 @@ NclQuark  varname;
                		 	if(ret < 0) {
                        			return(NhlWARNING);
                 		}
-				ret = nclfprintf(fp,"Total Size: %lld bytes\n",total * _NclSizeOf(thefile->file.var_info[i]->data_type));
+				ret = nclfprintf(fp,"Total Size: %lld bytes\n",(long long)total * _NclSizeOf(thefile->file.var_info[i]->data_type));
                 		if(ret < 0) {
                         		return(NhlWARNING);
                 		}
-				ret = nclfprintf(fp,"            %lld values\n",total);
+				ret = nclfprintf(fp,"            %lld values\n",(long long)total);
                 		if(ret < 0) {
                         		return(NhlWARNING);
                 		}
